домашка ресторан доработка

+++ src/RestaurantOrderingSystem/factory/MenuItemFactory.java
@@ -5,6 +5,6 @@ import RestaurantOrderingSystem.Enum.Category;
 import RestaurantOrderingSystem.model.MenuItem;
 
 public interface MenuItemFactory {
-    MenuItem createMenuItem(int id,String name, double price, boolean isAvailable);
+    MenuItem createMenuItem(Long id,String name, double price, boolean isAvailable);
     Category getCategory();
 }
